Infer angles or cell cycle phase based on gene expression data

Infer angles or cell cycle phase based on gene expression data

cycle_npreg_loglik(Y, sigma_est, funs_est, grids = 100,
  method.grid = c("pca", "uniform"))

Arguments

Y
Gene by sample expression matrix.

sigma_est
A vector of standard errors for each gene from the training data.

funs_est
A vector of cyclic functions estimated for each gene from the training data.

grids
number of bins to be selected along 0 to 2pi.

method.grid
The approach to initialize angles in the computation. uniform creates k equally-spaced bins (grids). pca uses gene expression values to infer angles, and then use these pca-based angles to move the cells to the closest bin (as defined by uniform).

Value

A list with the following three elements:

cell_times_est

Inferred angles or cell cycle phases, NOT ordered.

loglik_est

Log-likelihood estimates for each gene.

prob_per_cell_by_celltimes

Probabilities of each cell belong to each bin.