Predict test-sample ordering using training labels (no update)

Apply the estimates of cycle_npreg_insample to another gene expression dataset to infer an angle or cell cycle phase for each cell.

cycle_npreg_outsample(Y_test, sigma_est, funs_est,
  method.trend = c("trendfilter", "loess", "bspline"), polyorder = 2,
  method.grid = c("pca", "uniform"), ncores = 4,
  get_trend_estimates = F)

Arguments

Y_test
Gene expression data to be used for prediction. Gene by sample.

sigma_est
A vector of gene-specific standard error of the cyclic trends.

funs_est
A vector of cyclic functions estimating cyclic trends.

method.trend
Varous methods that can be applied to estimate cyclic trend of gene expression levels.

polyorder
We estimate cyclic trends of gene expression levels using nonparamtric trend filtering. The default fits second degree polynomials.

method.grid
Method for defining bins along the circle.

ncores
We use mclapply function for parallel computing.

get_trend_estimates
To re-estimate the cylic trend based on the predicted cell cycle phase or not (T or F). Default FALSE. This step calls trendfilter and is computationally intensive.

Value

A list with the following elements:

Y

The input gene expression marix.

cell_times_est

Inferred angles or cell cycle phases, NOT ordered.

loglik_est

Log-likelihood estimates for each gene.

cell_times_reordered

The inferred angles reordered (in ascending order).

Y_reorded

The input gene expression matrix reordered by cell_times_reordered.

sigma_reordered

Estimated standard error of the cyclic trend for each gene, reordered by cell_times_reordered.

funs_reordered

A list of functions for approximating the cyclic trends of gene express levels for each gene, reordered by cell_times_reordered.

mu_reordered

Estimated cyclic trend of gene expression values for each gene, reordered by cell_times_reordered.

prob_per_cell_by_celltimes

Probabilities of each cell belong to each bin.