Results of predicting cell cycle phase for samples from NA19098.

Pre-computed results. Applied cycle_npreg_outsample and results stored in fit_train to predict cell cycle phase for single-cell samples of NA19098.

fit_predict

Format

A list with the follwing elements

Y

a data.frame (gene by sample) of quantile-normailzed gene expression values

cell_times_est

a vector of estimated cell cycl phase values (range between 0 to 2pi)

loglik_est

a vector of estimated log-likelihoods for each cell

Y_reordered

Y ordered along columns by cell_times-est

cell_times_reordered

cell_times_est ordered from 0 to 2pi

funs_reordered

A lists of estimated cyclic functions for each gene using re-ordered gene expression values (rows of Y_reordered)

mu_reordered

A data.frame (gene by sample) of estimated expression for each gene across cells in the test data using funs_reordered

sigma_reordered

A vector of estimated standard error for each gene in the test data across cells using funs_reordered

prob_per_cell_by_celltimes

A data.frame (gene by sample) of log-likelihoods of of gene expression values in the test data

Y_reordered

Normalized expression values (log2CPM , )

cell_times_reordered

Estimated phase