Estimate parameters of the cyclic trends

This is used in both cycle_npreg_insample (training data fitting) and cycle_npreg_outsample (testing data prediction) to estimate cyclic trends of gene expression values. The function outputs for each gene standard error of the cyclic trend, cyclic function, and the estimated expression levels given the cyclic function.

cycle_npreg_mstep(
  Y,
  theta,
  method.trend = c("trendfilter", "loess", "bspline"),
  polyorder = 2,
  ncores = 2
)

Arguments

Y
Gene by sample expression matrix (log2CPM).

theta
Observed cell times.

method.trend
How to estimate cyclic trend of gene expression values? We offer three options: 'trendfilter' (fit_trendfilter()), 'loess' (fit_loess()) and 'bsplines' (fit_bspline()). 'trendfilter' provided the best fit in our study. But 'trendfilter` uses cross-validation and takes some time. Therefore, we recommend using bspline for quick results.

polyorder
We estimate cyclic trends of gene expression levels using nonparamtric trend filtering. The default fits second degree polynomials.

ncores
How many computing cores to use? We use doParallel package for parallel computing.

Value

A list with the following elements:

Y

Input gene expression data.

theta

Input angles.

mu_est

Estimated expression levels given the cyclic function for each gene.

sigma_est

Estimated standard error of the cyclic trends for each gene

funs

Estimated cyclic functions

See also